requirements.txt
numpy
scipy
pandas
matplotlib
yfinance
jupyter
wrds
pytest
python-dotenv